Moved EditorGradientBakerComponent.h into the Include directory so we can use it in tests
