+++ Gems/GradientSignal/Code/Include/GradientSignal/Editor/EditorGradientBakerComponent.h
