+++ Gems/GradientSignal/Code/gradientsignal_editor_files.cmake
@@ -7,6 +7,7 @@
 #
 
 set(FILES
+    Include/GradientSignal/Editor/EditorGradientBakerComponent.h
     Include/GradientSignal/Editor/EditorGradientComponentBase.h
     Include/GradientSignal/Editor/EditorGradientComponentBase.inl
     Include/GradientSignal/Editor/EditorGradientPreviewRenderer.h
@@ -20,7 +21,6 @@ set(FILES
     Source/Editor/EditorDitherGradientComponent.cpp
     Source/Editor/EditorDitherGradientComponent.h
     Source/Editor/EditorGradientBakerComponent.cpp
-    Source/Editor/EditorGradientBakerComponent.h
     Source/Editor/EditorGradientSurfaceDataComponent.cpp
     Source/Editor/EditorGradientSurfaceDataComponent.h
     Source/Editor/EditorGradientTransformComponent.cpp
